# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Recovy",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://recovy.herokuapp.com/api/v1"
    }
  ],
  "paths": {
...
    "/webinar": {
      "post": {
        "tags": ["Webinar"],
        "summary": "Membuat webinar baru",
        "description": "Membuat webinar baru memerlukan access token dari author",
        "parameters": [
          {
            "in": "header",
            "description": "Wajib mencamtumkan access token pada Header dengan format \"Authorization: bearer [token] \"",
            "name": "Authorization",
            "required": true,
            "schema": {
              "type": "string",
              "example": "bearer your_token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "authorID": {
                  
                    "type": "integer",
                    "minimum": 1
                  },
                  "picture": {
                   
                    "type": "string",
                    "maximum": 512
                  },
                  "title": {
                  
                    "type": "string",
                    "maximum": 255
                  },
                  "description": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string",
                  
                    "maximum": 128
                  },
                  "startDate": {
                    "type": "integer",
                  
                    "minimum": 1
                  },
                  "lastDate": {
                    "type": "integer",
                    
                    "minimum": 1
                  },
                  "time": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Jika sukses akan selalu mengembalikan object pada properti data",
            "content": {
              "application/json": {
                "schema": {
                 
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                         
                          "type": "integer",
                          "minimum": 1
                        },
                        "authorID": {
                         
                          "type": "integer",
                          "minimum": 1
                        },
                        "picture": {
                         
                          "type": "string",
                          "maximum": 512
                        },
                        "title": {
                        
                          "type": "string",
                          "maximum": 255
                        },
                        "description": {
                          "type": "string"
                        },
                        "category": {
                          "type": "string",
                 
                          "maximum": 128
                        },
                        "startDate": {
                          "type": "integer",
                         
                          "minimum": 1
                        },
                        "lastDate": {
                          "type": "integer",
                          
                          "minimum": 1
                        },
                        "time": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "integer",
                         
                          "minimum": 1
                        },
                        "updatedAt": {
                          "type": "integer",
                       
                          "minimum": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/webinar/{webinarID}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "description": "ID dari webinar",
            "required": true,
            "name": "webinarID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "tags": ["Webinar"],
        "summary": "Get webinar berdasar id-nya",
        "responses": {
          "200": {
            "description": "Jika sukses akan selalu mengembalikan object pada properti data",
            "content": {
              "application/json": {
                "schema": {
               
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
               
                          "type": "integer",
                          "minimum": 1
                        },
                        "authorID": {
                         
                          "type": "integer",
                          "minimum": 1
                        },
                        "picture": {
                         
                          "type": "string",
                          "maximum": 512
                        },
                        "title": {
                         
                          "type": "string",
                          "maximum": 255
                        },
                        "description": {
                          "type": "string"
                        },
                        "category": {
                          "type": "string",
                    
                          "maximum": 128
                        },
                        "startDate": {
                          "type": "integer",
                        
                          "minimum": 1
                        },
                        "lastDate": {
                          "type": "integer",
                      
                          "minimum": 1
                        },
                        "time": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "integer",
                       
                          "minimum": 1
                        },
                        "updatedAt": {
                          "type": "integer",
                         
                          "minimum": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Webinar"],
        "summary": "Menghapus webinar",
        "description": "Menghapus webinar memerlukan access token dari author atau pemilik. Jika menghapus item dari bukan miliknya akan mengembalikan error forbidden",
        "parameters": [
          {
            "in": "header",
            "description": "Wajib mencamtumkan access token pada Header dengan format \"Authorization: bearer [token] \"",
            "name": "Authorization",
            "required": true,
            "schema": {
              "type": "string",
              "example": "bearer your_token"
            }
          },
          {
            "required": true,
            "description": "ID dari webinar",
            "in": "path",
            "name": "webinarID",
            "schema":{
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jika sukses mengahpus, akan selalu mengembalikan properti success yang bernilai true",
            "content": {
              "application/json": {
                "schema": {
               
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/webinars/{category}": {
      "get": {
        "tags": ["Webinar"],
        "summary": "Mendapatkan kumpulan webinars berdasarkan category",
        "parameters": [
          {
            "in": "path",
            "name": "category",
            "example": "health",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Jika sukses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        
                        "properties": {
                          "id": {
                          
                            "type": "integer",
                            "minimum": 1
                          },
                          "author": {
                         
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "name": {
                                "type": "string"
                              },
                              "photo": {
                                "type": "string"
                              }
                            }
                          },
                          "picture": {
                            
                            "type": "string",
                            "maximum": 512
                          },
                          "title": {
                         
                            "type": "string",
                            "maximum": 255
                          },
                          "description": {
                            "type": "string"
                          },
                          "category": {
                            "type": "string",
                           
                            "maximum": 128
                          },
                          "startDate": {
                            "type": "integer",
                         
                            "minimum": 1
                          },
                          "lastDate": {
                            "type": "integer",
                          
                            "minimum": 1
                          },
                          "time": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "integer",
                           
                            "minimum": 1
                          },
                          "updatedAt": {
                            "type": "integer",
                          
                            "minimum": 1
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}